+++ tests/c_heat3dsc.c
@@ -40,13 +40,13 @@ const double threshold = 1.0e-15; // should be small enough to catch any obvious
 
 double SourceQ_t(double x,double y,double z,double A_x,double B_y,double C_z,double k_0)
 {
-  double Q_T = k_0 * cosl(A_x * x) * cosl(B_y * y) * cosl(C_z * z) * (A_x * A_x + B_y * B_y + C_z * C_z);
+  double Q_T = k_0 * cos(A_x * x) * cos(B_y * y) * cos(C_z * z) * (A_x * A_x + B_y * B_y + C_z * C_z);
   return Q_T;
 }
 
 double Source_exact_t(double x,double y,double z,double A_x,double B_y,double C_z)
 {
-  double exact_t = cosl(A_x * x) * cosl(B_y * y) * cosl(C_z * z);
+  double exact_t = cos(A_x * x) * cos(B_y * y) * cos(C_z * z);
   return exact_t;
 }
